LPS-33872 Implement roles method in the base so the membership policy classes doesn't need to implement those methods in case the policy it's not related to roles at all
