+++ portal-service/src/com/liferay/portal/security/membershippolicy/DummySiteMembershipPolicy.java
@@ -16,8 +16,6 @@ package com.liferay.portal.security.membershippolicy;
 
 import com.liferay.portal.kernel.util.UnicodeProperties;
 import com.liferay.portal.model.Group;
-import com.liferay.portal.model.Role;
-import com.liferay.portal.model.UserGroupRole;
 import com.liferay.portlet.asset.model.AssetCategory;
 import com.liferay.portlet.asset.model.AssetTag;
 
@@ -34,11 +32,6 @@ public class DummySiteMembershipPolicy extends BaseSiteMembershipPolicy {
 
 	public void checkMembership(
 		long[] userIds, long[] addGroupIds, long[] removeGroupIds) {
-	}
-
-	public void checkRoles(
-		List<UserGroupRole> addUserGroupRoles,
-		List<UserGroupRole> removeUserGroupRoles) {
 	}
 
 	@Override
@@ -65,11 +58,6 @@ public class DummySiteMembershipPolicy extends BaseSiteMembershipPolicy {
 		long[] userIds, long[] addGroupIds, long[] removeGroupIds) {
 	}
 
-	public void propagateRoles(
-		List<UserGroupRole> addUserGroupRoles,
-		List<UserGroupRole> removeUserGroupRoles) {
-	}
-
 	@Override
 	public void verifyPolicy(Group group) {
 	}
@@ -79,14 +67,6 @@ public class DummySiteMembershipPolicy extends BaseSiteMembershipPolicy {
 		List<AssetTag> oldAssetTags,
 		Map<String, Serializable> oldExpandoAttributes,
 		UnicodeProperties oldTypeSettingsProperties) {
-	}
-
-	public void verifyPolicy(Role role) {
-	}
-
-	public void verifyPolicy(
-		Role role, Role oldRole,
-		Map<String, Serializable> oldExpandoAttributes) {
 	}
 
 }
